+++ src/native/L2NormSetCover.h
@@ -25,9 +25,9 @@ private:
     // Computed
     vector<unique_ptr<Term>> allTerms;
 
-    unordered_map<long long, long double> idToRedness;
-    unordered_map<const Term *, long double> termToAverageRedness;
-    vector<std::pair<const Term *, long double>> sortedTermAndAverageRedness;
+    unordered_map<long long, double> idToRedness;
+    unordered_map<const Term *, double> termToAverageRedness;
+    vector<std::pair<const Term *, double>> sortedTermAndAverageRedness;
 
     // Given
     unique_ptr<Dataset> dataset;
@@ -38,13 +38,13 @@ private:
     void createTermsForIndices(const Dataset *dataset, vector<unique_ptr<Term>> &allTerms, long long i1, long long i2);
 
     // Helpers (L2NormSetCover run)
-    static pair<long double, long double> calculateCoefficients(unique_ptr<Dataset> &dataset, Work &work);
-    static bool coefficientsValid(long double a1, long double a2);
-    void deriveRednessMaps(vector<long long> &yDimensions, long double a1, long double a2);
+    static pair<double, double> calculateCoefficients(unique_ptr<Dataset> &dataset, Work &work);
+    static bool coefficientsValid(double a1, double a2);
+    void deriveRednessMaps(vector<long long> &yDimensions, double a1, double a2);
     dynamic_bitset<> bitsetUnion(vector<const Term *> &terms);
-    long double rednessOfTerm(const Term *term);
-    long double rednessOfTerms(vector<const Term *> &terms);
-    long double errorRate(vector<const Term *> &terms);
+    double rednessOfTerm(const Term *term);
+    double rednessOfTerms(vector<const Term *> &terms);
+    double errorRate(vector<const Term *> &terms);
 public:
     L2NormSetCover(unique_ptr<Dataset> dataset, int dnfSize, double mu);
     Result run(Work &work);
